# .services/artex-platform-api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: ARTEX Platform API
  version: 1.0.0
  description: |
    Public API for managing ARTEX living art projects.

    Provides full CRUD operations on projects (ConfigJson, StateJson, ProjectPackageData),
    real-time state subscriptions via WebSocket, and asset management.

    ## Authentication

    All endpoints require a Bearer token in the `Authorization` header.

    Phase 1 uses API keys (server-to-server). Phase 2 adds Firebase Auth JWT tokens
    for user-context operations. Phase 3 introduces OAuth2 for full third-party flows.

    ## Rate Limiting

    Default: 100 requests/minute per API key. WebSocket connections: 5 concurrent per key.
    State push via WebSocket: 30 frames/second max.

    ## Real-Time Updates

    Connect to `wss://{host}/v1/ws/projects/{projectId}/state` for live state streaming.
    See the WebSocket Protocol section for frame format.

    ## Future: Extension Gateway (Option C)

    A future `/v1/extensions/*` namespace is reserved for remote extension registration
    (shaders, media inputs). This will allow external apps to push live signals
    into running experiences without being bundled in the same runtime.

  contact:
    name: CODAME
    url: https://codame.com
  license:
    name: Proprietary
    identifier: LicenseRef-CODAME
servers:
  - url: https://api.artex.live/v1
    description: Production
  - url: https://api-staging.artex.live/v1
    description: Staging
  - url: http://localhost:8080/v1
...
security:
  - BearerAuth: []
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [Projects]
      summary: List all projects accessible to the caller
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: status
          in: query
          description: Filter by project lifecycle status
          schema:
            type: string
            enum: [draft, running, stopped, archived]
      responses:
        "200":
          description: Paginated project list
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProjectSummary"
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer

    post:
      operationId: createProject
      tags: [Projects]
      summary: Create a new project
      description: |
        Creates a new ARTEX project with the provided configuration.
        A default StateJson is generated from the config using `createStateFromConfig()`.
        The project starts in `draft` status.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/ValidationError"
        "422":
          $ref: "#/components/responses/ContractError"
  /projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/projectId"

    get:
      operationId: getProject
      tags: [Projects]
      summary: Get full project details
      description: Returns config, state, package data, and metadata.
      responses:
        "200":
          description: Full project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteProject
      tags: [Projects]
      summary: Delete a project
      description: Permanently deletes the project and all associated assets.
      responses:
        "204":
          description: Project deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
  /projects/{projectId}/run:
    parameters:
      - $ref: "#/components/parameters/projectId"

    post:
      operationId: runProject
      tags: [Lifecycle]
      summary: Start or resume a project
      description: |
        Transitions the project to `running` status. If already running, this is a no-op.
        Initializes state from config if no state exists.
        Opens the project for real-time WebSocket subscriptions.
      responses:
        "200":
          description: Project is now running
          content:
            application/json:
              schema:
                type: object
                properties:
                  projectId:
                    type: string
                  status:
                    type: string
                    enum: [running]
                  wsUrl:
                    type: string
                    description: WebSocket URL for real-time state subscription
  /projects/{projectId}/stop:
    parameters:
      - $ref: "#/components/parameters/projectId"

    post:
      operationId: stopProject
      tags: [Lifecycle]
      summary: Stop a running project
      description: |
        Transitions to `stopped`. Disconnects all WebSocket subscribers
        with a `project_stopped` close frame.
        State is preserved and can be resumed with `/run`.
      responses:
        "200":
          description: Project stopped
          content:
            application/json:
              schema:
                type: object
                properties:
                  projectId:
                    type: string
                  status:
                    type: string
                    enum: [stopped]
  /projects/{projectId}/reset:
    parameters:
      - $ref: "#/components/parameters/projectId"

    post:
      operationId: resetProject
      tags: [Lifecycle]
      summary: Reset project state to initial
      description: |
        Regenerates StateJson from the current ConfigJson using
        `createStateFromConfig()`. Broadcasts `state_replaced` if running.
      responses:
        "200":
          description: State reset to initial
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StateJson"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Phase 1: API key as Bearer token.
        Phase 2: Firebase Auth JWT.
        Phase 3: OAuth2 access token.
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      description: Unique project identifier
      schema:
        type: string
  responses:
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ContractError:
      description: Data violates ARTEX contract rules
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error context
    ProjectSummary:
      type: object
      required: [projectId, title, status, createdAt, updatedAt]
      properties:
        projectId:
          type: string
        title:
          type: string
        artistName:
          type: string
        status:
          type: string
          enum: [draft, running, stopped, archived]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        activeSubscribers:
          type: integer
          description: Number of active WebSocket connections
    Project:
      type: object
      required: [projectId, status, config, state, createdAt, updatedAt]
      properties:
        projectId:
          type: string
        status:
          type: string
          enum: [draft, running, stopped, archived]
        config:
          $ref: "#/components/schemas/ConfigJson"
        state:
          $ref: "#/components/schemas/StateJson"
        packageData:
          $ref: "#/components/schemas/ProjectPackageData"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateProjectRequest:
      type: object
      required: [config]
      properties:
        config:
          $ref: "#/components/schemas/ConfigJson"
        packageData:
          $ref: "#/components/schemas/ProjectPackageData"
    ConfigJson:
      type: object
      description: Full ARTEX artwork configuration. See @artex/contract ConfigJson.
      required: [version, title, story, layers, animation, evolution, interaction, constraints]
      properties:
        version:
          type: integer
        title:
          type: string
        artistName:
          type: string
        story:
          type: string
        # ... remaining fields match @artex/contract ConfigJson exactly
        # Full schema generated from TypeScript types at build time
    StateJson:
      type: object
      description: Runtime state. See @artex/contract StateJson.
      required: [installTimestamp, randomSeed, timeOffsetSeconds, currentPhaseLabel, parameters, eventsLog]
      properties:
        artworkId:
          type: string
        installTimestamp:
          type: string
          format: date-time
        randomSeed:
          type: integer
        timeOffsetSeconds:
          type: number
        currentPhaseLabel:
          type: string
        parameters:
          type: object
          properties:
            breathingIntensity:
              type: number
            colorTemperature:
              type: number
            parallaxShift:
              type: number
            noiseIntensity:
              type: number
        eventsLog:
          type: array
          items:
            type: object
            properties:
              t:
                type: number
              event:
                type: string
    ProjectPackageData:
      type: object
      description: Extended project data. See @artex/contract ProjectPackageData.
      required: [version, shader, shaderChannels]
      properties:
        version:
          type: integer
        shader:
          type: object
          description: ProjectShaderState
        shaderChannels:
          type: object
          properties:
            paths:
              type: array
              items:
                type: string
                nullable: true
            sources:
              type: array
              items:
                type: string
                enum: [none, file, camera]
        runtime:
          type: object
        ai:
          type: object
        experimental:
          type: object
...